voice-micro-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
python-dotenv==1.0.0
googletrans==4.0.0rc1
google-generativeai==0.3.2
gspread==5.12.0
google-auth==2.23.4
python-multipart==0.0.6